Код разделён на модули,точка входа подключена в html файл
